Organize and polish stories
